requirements.txt
Django~=5.1.1
django-seed~=0.3.1
psycopg2-binary~=2.9.9
django-debug-toolbar~=4.4.6
django-bootstrap5~=24.3
django-bootstrap-icons~=0.8.6
django-environ~=0.11.2
djangorestframework~=3.14.0
requests~=2.31.0